コードのセーブとロードはできる。ステージはまだ。

+++ Item.cs
@@ -5,7 +5,7 @@ namespace HackTheWorld
 {
     class Item : GameObject
     {
-        public Item(int x, int y) : base(x, y)
+        public Item(float x, float y) : base(x, y)
         {
             X += CellSize / 4;
             Y += CellSize / 2;
@@ -14,12 +14,12 @@ namespace HackTheWorld
             ObjectType = ObjectType.Item;
         }
 
-        public Item(int x, int y, int vx, int vy) : base(x, y, vx, vy)
+        public Item(float x, float y, float vx, float vy) : base(x, y, vx, vy)
         {
             ObjectType = ObjectType.Item;
         }
 
-        public Item(int x, int y, int vx, int vy, int w, int h) : base(x, y, vx, vy, w, h)
+        public Item(float x, float y, float vx, float vy, float w, float h) : base(x, y, vx, vy, w, h)
         {
             ObjectType = ObjectType.Item;
         }
